requirements.txt
# Docker integration for containerized deployment
docker==4.41.2

# Core dependencies for Kafka-based data streaming
kafka-python==3.8
confluent-kafka==2.2.0  # Kafka CLI tools
requests==2.31.0
jsonschema==4.21.0

# CSV and data processing utilities
pandas==2.1.0
numpy==1.25.0
openpyxl==3.1.2

# Web automation with Playwright
playwright==1.41.0

# Time & regex handling
python-dateutil==2.9.0
regex==2023.12.1

# Databricks utilities for data integration
databricks-cli==0.17.7
delta-spark==3.0.0
pyspark==3.4.1

# Machine Learning tools within Databricks
tensorflow==2.15.0
scikit-learn==1.3.0

# Apache Airflow for pipeline orchestration
apache-airflow==2.9.0
apache-airflow-providers-databricks==3.4.0
